services/easyocrcb_app/requirements.txt
catboost==1.2.3
easyocr==1.7.1
Flask==3.0.3
Pillow==10.3.0
